requirements.txt
Flask==2.1.1
Pillow==8.4.0